Feature leadmagnet update

+++ components/slide-customizer.tsx
@@ -13,12 +13,19 @@ import {
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from "@/components/ui/tooltip"
+} from "@/components/ui/tooltip";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
 } from "@/components/ui/popover";
+
+import { Input } from "@/components/ui/input";
+import { Button } from "@/components/ui/button";
+import { categories, categoriesImages } from "@/data";
+import { useDebounce } from "use-debounce";
+import { Skeleton } from "@/components/ui/skeleton";
+import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area";
 
 const SlideCustomizer = ({
   slidesStyles,
@@ -31,15 +38,17 @@ const SlideCustomizer = ({
   updateStyleSlides: any;
   editor: any;
 }) => {
+  const perPageImages = 10;
   const defaultTextColor = { r: 0, g: 0, b: 0, a: 1 };
   const defaultBgColor = { r: 241, g: 245, b: 249, a: 1 };
   const [slideStyle, setSlideStyle] = useState<SlideStyle | undefined>();
   const [textColor, setTextColor] = useState(defaultTextColor);
   const [bgColor, setBgColor] = useState(defaultBgColor);
-  const [showTextColorPicker, setShowTextColorPicker] = useState(false);
-  const [showBgColorPicker, setShowBgColorPicker] = useState(false);
   const [image, setImage] = useState("");
   const [showResetBtn, setShowResetBtn] = useState(false);
+  const [search, setSearch] = useState("");
+  const [category, setCategory] = useState(categories[0]);
+  const [images, setImages] = useState([]);
   const imageRef = useRef(null);
   const componentRef = useRef(null);
   const presetColors = [
@@ -189,6 +198,9 @@ const SlideCustomizer = ({
     },
   ];
 
+  // debaounce search after 3 seconds
+  const [debouncedSearch] = useDebounce(search, 2000);
+
   // arrange original data from databse
   useEffect(() => {
     // @ts-ignore
@@ -225,7 +237,7 @@ const SlideCustomizer = ({
   }
 
   async function handleImageUpload(e: any) {
-    // delete image from vercel blob if exists
+    // delete image from vercel blob if exists and image is uploaded to vercel blob
     if (image) {
       await deleteImage();
     }
@@ -251,15 +263,16 @@ const SlideCustomizer = ({
   }
 
   async function deleteImage() {
-    const response = await fetch("/api/upload", {
-      method: "DELETE",
-      body: JSON.stringify({ image }),
-    });
-    const { success } = await response.json();
-    if (success) {
-      setImage("");
-      handleValueChange("image", "");
+    if (image.includes("blob.vercel-storage.com")) {
+      const response = await fetch("/api/upload", {
+        method: "DELETE",
+        body: JSON.stringify({ image }),
+      });
+      const { success } = await response.json();
     }
+
+    setImage("");
+    handleValueChange("image", "");
   }
 
   function changeEditorTextColor(type: string, value: any) {
@@ -352,46 +365,81 @@ const SlideCustomizer = ({
     }
   }, [textColor, bgColor, image]);
 
-  // useEffect(() => {
-  //   document.addEventListener("mousedown", handleOutsideClick);
-  //   return () => {
-  //     document.removeEventListener("mousedown", handleOutsideClick);
-  //   };
-  // });
+  // fetching images from unsplash api
+  async function fetchImageFromUnsplash() {
+    const request = await fetch(
+      `https://api.unsplash.com/search/photos?query=${debouncedSearch}&per_page=${perPageImages}`,
+      {
+        headers: {
+          Authorization: `Client-ID ${process.env.NEXT_PUBLIC_UNSPLASH_ACCESS_KEY}`,
+        },
+      },
+    );
+    const images = await request.json();
+    setImages(images.results);
+  }
 
-  // const handleOutsideClick = (e: any) => {
-  //   // @ts-ignore
-  //   if (componentRef.current && !componentRef.current.contains(e.target)) {
-  //     setShowTextColorPicker(false);
-  //     setShowBgColorPicker(false);
-  //   }
-  // };
+  // search images from unsplash when user search
+  useEffect(() => {
+    let timeout: any = null;
+    if (debouncedSearch) {
+      setLoading(true);
+      fetchImageFromUnsplash();
+      timeout = setTimeout(() => {
+        setLoading(false);
+      }, 500);
+    } else {
+      timeout = setTimeout(() => {
+        setLoading(false);
+      }, 500);
+      setCategory("gradients");
+    }
+
+    return () => {
+      clearTimeout(timeout);
+    };
+  }, [debouncedSearch]);
+
+  // show image when category change
+  useEffect(() => {
+    if (category) {
+      const data = categoriesImages.find((item) => item.category === category);
+
+      // @ts-ignore
+      setImages(data.images ?? []);
+    } else {
+      setImages([]);
+    }
+  }, [category]);
+
+  const [loading, setLoading] = useState(true);
 
   return (
     <>
       {/* text color picker wrapper */}
 
       <Popover>
+        <TooltipProvider delayDuration={100}>
+          <Tooltip>
+            <TooltipTrigger asChild>
+              <PopoverTrigger
+                ref={componentRef}
+              
+                className="shadow-sm absolute bottom-2 right-2 flex  items-center justify-center  opacity-0  transition ease-in-out group-hover:opacity-100"
+              >
+                <Button size="icon" variant="secondary">
+                <Palette strokeWidth={"1.5px"} width={20} />
+                </Button>
+               
+              </PopoverTrigger>
+            </TooltipTrigger>
+            <TooltipContent side="bottom">
+              <p>Customize design</p>
+            </TooltipContent>
+          </Tooltip>
+        </TooltipProvider>
 
-      <TooltipProvider delayDuration={100}>
-      <Tooltip>
-        <TooltipTrigger asChild>
-        <PopoverTrigger
-          ref={componentRef}
-          className="opacity-0 group-hover:opacity-100 transition ease-in-out absolute bottom-5 right-5 flex h-10 w-10 items-center justify-center rounded-full bg-white p-2 shadow-sm"
-        >
-          <Palette strokeWidth={"1.5px"} width={20} />
-        </PopoverTrigger>
-        </TooltipTrigger>
-        <TooltipContent side="bottom">
-          <p>Customize design</p>
-        </TooltipContent>
-      </Tooltip>
-    </TooltipProvider>
-
-
-       
-        <PopoverContent className="mt-2 w-auto rounded-full">
+        <PopoverContent className="mt-2 w-auto rounded-lg shadow-lg border-0 py-2">
           <div className="flex items-center gap-x-2">
             <div className="relative flex items-center gap-x-2">
               <p className="text-xs text-gray-500">TEXT</p>
@@ -451,19 +499,111 @@ const SlideCustomizer = ({
                 </PopoverContent>
               </Popover>
 
-              <button
-                type="button"
-                className="text-gray-500 hover:text-gray-600"
-                onClick={openImageDialog}
-              >
-                <ImageIcon strokeWidth={"1.5px"} width={20} />
-              </button>
-              <input
-                type="file"
-                ref={imageRef}
-                onChange={handleImageUpload}
-                className="hidden"
-              />
+              <Popover>
+                <PopoverTrigger>
+                
+                    <ImageIcon className="text-gray-500" strokeWidth={"1.5px"} width={20} />
+                
+                </PopoverTrigger>
+                <PopoverContent className="w-min rounded-xl shadow-2xl border-0 my-4">
+                  <div className="w-[250px]">
+                    <Input
+                      type="text"
+                      placeholder="Search"
+                      value={search}
+                      onChange={(e: any) => setSearch(e.target.value as string)}
+                    />
+                    {/* showing categories */}
+                    <div className="mt-3 flex flex-wrap items-center gap-2">
+                      {categories.map((item: string) => (
+                        <Button
+                          variant={category === item ? "default" : "outline"}
+                          size="xs"
+                          className="capitalize basis-1/4 max-w-[22%]"
+                          key={item}
+                          onClick={() =>
+                            category === item
+                              ? setCategory("")
+                              : setCategory(item)
+                          }
+                        >
+                          {item}
+                        </Button>
+                      ))}
+                    </div>
+
+                    {/* showing images */}
+                    <ScrollArea className="h-[200px] w-[250px]">
+                      <div>
+                        <div className="grid grid-cols-2 gap-4">
+                          {images.map((item: any, idx: number) => (
+                            <div className="group relative mb-3" key={idx}>
+                              {loading ? (
+                               
+                                    <Skeleton className="h-[30px] w-full rounded-full" />
+                                    
+                               
+                              
+                              ) : (
+                                <Image
+                                  className="h-auto rounded-xl w-full cursor-pointer object-contain"
+                                  width={120}
+                                  height={60}
+                                  src={item.urls.regular}
+                                  alt="background image"
+                                  onClick={() =>
+                                    handleValueChange(
+                                      "image",
+                                      item.urls.regular as string,
+                                    )
+                                  }
+                                />
+                              )}
+                              {item.user && (
+                                <div className="absolute left-0 bottom-0 inset-0 flex items-end justify-center p-1 opacity-0 transition-opacity group-hover:opacity-100">
+                                  <small className="text-[8px] text-white">
+                                    Photo by{" "}
+                                    <a
+                                      href={item.user?.links?.self}
+                                      className="text-blue-400 hover:text-blue-300"
+                                    >
+                                      {item.user?.name}
+                                    </a>{" "}
+                                    on{" "}
+                                    <a
+                                      href="https://unsplash.com"
+                                      className="text-blue-400 hover:text-blue-300"
+                                    >
+                                      Unsplash
+                                    </a>
+                                  </small>
+                                </div>
+                              )}
+                            </div>
+                          ))}
+                        </div>
+                      </div>
+                    </ScrollArea>
+                    <div className="mt-3">
+                      <Button
+                        type="button"
+                        variant="secondary"
+                        className="w-full"
+                        onClick={openImageDialog}
+                      >
+                        Upload Image
+                      </Button>
+                      <input
+                        type="file"
+                        ref={imageRef}
+                        onChange={handleImageUpload}
+                        className="hidden"
+                      />
+                    </div>
+                  </div>
+                </PopoverContent>
+              </Popover>
+
               {image && (
                 <>
                   {/* <Image
